[hotfix][runtime] Extract SlotManager interface
Rename SlotManager -> SlotManagerImpl and let it implement the SlotManager interface
