+++ flink-runtime/src/test/java/org/apache/flink/runtime/resourcemanager/slotmanager/SlotManagerTest.java
@@ -95,7 +95,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 /**
- * Tests for the {@link SlotManager}.
+ * Tests for the {@link SlotManagerImpl}.
  */
 public class SlotManagerTest extends TestLogger {
 
@@ -118,7 +118,7 @@ public class SlotManagerTest extends TestLogger {
 		final SlotStatus slotStatus2 = new SlotStatus(slotId2, resourceProfile);
 		final SlotReport slotReport = new SlotReport(Arrays.asList(slotStatus1, slotStatus2));
 
-		try (SlotManager slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
+		try (SlotManagerImpl slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
 			slotManager.registerTaskManager(taskManagerConnection, slotReport);
 
 			assertTrue("The number registered slots does not equal the expected number.", 2 == slotManager.getNumberRegisteredSlots());
@@ -162,7 +162,7 @@ public class SlotManagerTest extends TestLogger {
 			resourceProfile,
 			"foobar");
 
-		try (SlotManager slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
+		try (SlotManagerImpl slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
 			slotManager.registerTaskManager(taskManagerConnection, slotReport);
 
 			assertTrue("The number registered slots does not equal the expected number.", 2 == slotManager.getNumberRegisteredSlots());
@@ -265,7 +265,7 @@ public class SlotManagerTest extends TestLogger {
 
 		ResourceActions resourceManagerActions = new TestingResourceActionsBuilder().build();
 
-		try (SlotManager slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
+		try (SlotManagerImpl slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
 			final CompletableFuture<Tuple5<SlotID, JobID, AllocationID, String, ResourceManagerId>> requestFuture = new CompletableFuture<>();
 			// accept an incoming slot request
 			final TaskExecutorGateway taskExecutorGateway = new TestingTaskExecutorGatewayBuilder()
@@ -318,7 +318,7 @@ public class SlotManagerTest extends TestLogger {
 
 		final TaskExecutorConnection taskManagerConnection = new TaskExecutorConnection(resourceID, taskExecutorGateway);
 
-		try (SlotManager slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
+		try (SlotManagerImpl slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
 			slotManager.registerTaskManager(taskManagerConnection, slotReport);
 
 			TaskManagerSlot slot = slotManager.getSlot(slotId);
@@ -375,7 +375,7 @@ public class SlotManagerTest extends TestLogger {
 		final SlotStatus slotStatus = new SlotStatus(slotId, resourceProfile);
 		final SlotReport slotReport = new SlotReport(slotStatus);
 
-		try (SlotManager slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
+		try (SlotManagerImpl slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
 
 			assertTrue("The slot request should be accepted", slotManager.registerSlotRequest(slotRequest));
 
@@ -415,7 +415,7 @@ public class SlotManagerTest extends TestLogger {
 		final SlotStatus slotStatus = new SlotStatus(slotId, resourceProfile, jobId, allocationId);
 		final SlotReport slotReport = new SlotReport(slotStatus);
 
-		try (SlotManager slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
+		try (SlotManagerImpl slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
 
 			slotManager.registerTaskManager(
 				taskExecutorConnection,
@@ -521,7 +521,7 @@ public class SlotManagerTest extends TestLogger {
 		final SlotStatus slotStatus = new SlotStatus(slotId, resourceProfile1);
 		final SlotReport slotReport = new SlotReport(slotStatus);
 
-		try (SlotManager slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
+		try (SlotManagerImpl slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
 			slotManager.registerTaskManager(taskManagerConnection, slotReport);
 			assertTrue(slotManager.registerSlotRequest(slotRequest1));
 
@@ -563,7 +563,7 @@ public class SlotManagerTest extends TestLogger {
 		final SlotStatus slotStatus = new SlotStatus(slotId, new ResourceProfile(2.0, 2));
 		final SlotReport slotReport = new SlotReport(slotStatus);
 
-		try (SlotManager slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
+		try (SlotManagerImpl slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
 			slotManager.registerTaskManager(taskManagerConnection, slotReport);
 			assertTrue(slotManager.registerSlotRequest(slotRequest1));
 
@@ -641,7 +641,7 @@ public class SlotManagerTest extends TestLogger {
 		final TaskExecutorGateway taskExecutorGateway = mock(TaskExecutorGateway.class);
 		final TaskExecutorConnection taskManagerConnection = new TaskExecutorConnection(resourceId, taskExecutorGateway);
 
-		try (SlotManager slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
+		try (SlotManagerImpl slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
 			// check that we don't have any slots registered
 			assertTrue(0 == slotManager.getNumberRegisteredSlots());
 
@@ -747,7 +747,7 @@ public class SlotManagerTest extends TestLogger {
 		final SlotStatus slotStatus2 = new SlotStatus(slotId2, resourceProfile);
 		final SlotReport slotReport = new SlotReport(Arrays.asList(slotStatus1, slotStatus2));
 
-		try (SlotManager slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
+		try (SlotManagerImpl slotManager = createSlotManager(resourceManagerId, resourceManagerActions)) {
 
 			slotManager.registerTaskManager(taskManagerConnection, slotReport);
 
@@ -827,7 +827,7 @@ public class SlotManagerTest extends TestLogger {
 
 		final Executor mainThreadExecutor = TestingUtils.defaultExecutor();
 
-		try (final SlotManager slotManager = SlotManagerBuilder.newBuilder().build()) {
+		try (final SlotManagerImpl slotManager = SlotManagerBuilder.newBuilder().build()) {
 
 			slotManager.start(resourceManagerId, mainThreadExecutor, resourceManagerActions);
 
@@ -943,7 +943,7 @@ public class SlotManagerTest extends TestLogger {
 
 		final Executor mainThreadExecutor = TestingUtils.defaultExecutor();
 
-		try (final SlotManager slotManager = SlotManagerBuilder.newBuilder()
+		try (final SlotManagerImpl slotManager = SlotManagerBuilder.newBuilder()
 			.setTaskManagerTimeout(Time.of(taskManagerTimeout, TimeUnit.MILLISECONDS))
 			.build()) {
 
@@ -1046,7 +1046,7 @@ public class SlotManagerTest extends TestLogger {
 		final TestingTaskExecutorGateway taskExecutorGateway = new TestingTaskExecutorGatewayBuilder().createTestingTaskExecutorGateway();
 		final TaskExecutorConnection taskExecutorConnection = new TaskExecutorConnection(taskManagerId, taskExecutorGateway);
 
-		try (final SlotManager slotManager = SlotManagerBuilder.newBuilder().build()) {
+		try (final SlotManagerImpl slotManager = SlotManagerBuilder.newBuilder().build()) {
 
 			slotManager.start(ResourceManagerId.generate(), Executors.directExecutor(), resourceActions);
 
@@ -1095,7 +1095,7 @@ public class SlotManagerTest extends TestLogger {
 	 */
 	@Test
 	public void testSlotRequestFailure() throws Exception {
-		try (final SlotManager slotManager = createSlotManager(ResourceManagerId.generate(),
+		try (final SlotManagerImpl slotManager = createSlotManager(ResourceManagerId.generate(),
 			new TestingResourceActionsBuilder().build())) {
 
 			final SlotRequest slotRequest = new SlotRequest(new JobID(), new AllocationID(), ResourceProfile.UNKNOWN, "foobar");
@@ -1150,7 +1150,7 @@ public class SlotManagerTest extends TestLogger {
 	 */
 	@Test
 	public void testSlotRequestRemovedIfTMReportAllocation() throws Exception {
-		try (final SlotManager slotManager = createSlotManager(ResourceManagerId.generate(),
+		try (final SlotManagerImpl slotManager = createSlotManager(ResourceManagerId.generate(),
 				new TestingResourceActionsBuilder().build())) {
 
 			final JobID jobID = new JobID();
@@ -1326,8 +1326,8 @@ public class SlotManagerTest extends TestLogger {
 		return new SlotRequest(jobId, new AllocationID(), resourceProfile, "foobar1");
 	}
 
-	private SlotManager createSlotManager(ResourceManagerId resourceManagerId, ResourceActions resourceManagerActions) {
-		SlotManager slotManager = SlotManagerBuilder.newBuilder().build();
+	private SlotManagerImpl createSlotManager(ResourceManagerId resourceManagerId, ResourceActions resourceManagerActions) {
+		SlotManagerImpl slotManager = SlotManagerBuilder.newBuilder().build();
 		slotManager.start(resourceManagerId, Executors.directExecutor(), resourceManagerActions);
 		return slotManager;
 	}
@@ -1348,7 +1348,7 @@ public class SlotManagerTest extends TestLogger {
 				}))
 			.build();
 
-		try (final SlotManager slotManager = createSlotManager(
+		try (final SlotManagerImpl slotManager = createSlotManager(
 			ResourceManagerId.generate(),
 			testingResourceActions)) {
 
@@ -1377,7 +1377,7 @@ public class SlotManagerTest extends TestLogger {
 		final TestingResourceActions resourceActions = new TestingResourceActionsBuilder()
 			.setAllocateResourceFunction(convert(value -> numberSlots))
 			.build();
-		try (final SlotManager slotManager = createSlotManager(ResourceManagerId.generate(), resourceActions)) {
+		try (final SlotManagerImpl slotManager = createSlotManager(ResourceManagerId.generate(), resourceActions)) {
 			final JobID jobId = new JobID();
 
 			final SlotRequest slotRequest = createSlotRequest(jobId);
@@ -1403,7 +1403,7 @@ public class SlotManagerTest extends TestLogger {
 			.setAllocateResourceFunction(convert(value -> numberSlots))
 			.build();
 
-		try (final SlotManager slotManager = createSlotManager(ResourceManagerId.generate(), resourceActions)) {
+		try (final SlotManagerImpl slotManager = createSlotManager(ResourceManagerId.generate(), resourceActions)) {
 			final JobID jobId = new JobID();
 			assertThat(slotManager.registerSlotRequest(createSlotRequest(jobId)), is(true));
 
@@ -1438,7 +1438,7 @@ public class SlotManagerTest extends TestLogger {
 			.setAllocateResourceFunction(convert(value -> numberSlots))
 			.build();
 
-		try (final SlotManager slotManager = createSlotManager(ResourceManagerId.generate(), resourceActions)) {
+		try (final SlotManagerImpl slotManager = createSlotManager(ResourceManagerId.generate(), resourceActions)) {
 			final JobID jobId = new JobID();
 			final ResourceProfile requestedSlotProfile = new ResourceProfile(1.0, 1);
 
@@ -1469,7 +1469,7 @@ public class SlotManagerTest extends TestLogger {
 			.setAllocateResourceFunction(convert(value -> numberSlots))
 			.build();
 
-		try (final SlotManager slotManager = createSlotManager(ResourceManagerId.generate(), resourceActions)) {
+		try (final SlotManagerImpl slotManager = createSlotManager(ResourceManagerId.generate(), resourceActions)) {
 			final JobID jobId = new JobID();
 			assertThat(slotManager.registerSlotRequest(createSlotRequest(jobId)), is(true));
